Reverse GPX/EXIF file-loading logic
The most common workflow will most likely be loading a large number of photos, and then loading only one or two GPX files. So the code now assumes that files are photos, and if the photo loader fails to load, then it assumes they're GPX, and failing that it shows an error.