Champlain overhaul, plus lots of random changes
Obviously I'm terrible at this 'git' thing, because this commit includes quite a bit of code style cleanup, the addition of getting photo timestamps from pyexiv2 rather than from the filesystem, the totally brand-new addition of ChamplainPolygon for showing the loaded GPX data (yeah that's right, I wrote the GPX parser forever ago, and the program only now has a way of displaying the data). Also, I call self.map_view.center_on() for each GPX point loaded, so you get to watch as the program follows the loading of the GPX data. It's tedious for large files but amusing for small ones.  Also increased the default size of the window. Oh, and and! I also added self.PHOTO_MARKER column to the liststore, so that when you delete photos, it removes the ChamplainMarker from the ChamplainView.