Added Statusbar as a wrapper around the progressbar
Initially I really liked the idea of a progressbar that was only visible when the program is processing data, but what was happening was that self.progressbar.show() was shrinking the visible area of the map, and the map is set to maintain it's center point when it's viewable area changes, which was causing the map to "jitter" every time the progressbar was shown or hidden. Having a constant statusbar eliminates that problem and makes the map look much more stable, even if I generally dislike the vertical-space-wastage of a statusbar (widescreen, etc).