Eliminate the "jump to" button in favor of doing the comparison automatically
With this commit, each photo is immediately compared to the loaded GPX data upon successful loading of the photo, and in the event that you load your photos before loading the GPX, it also does the comparison at the end of each load of GPX data. Still allows manual override with the 'apply' button.