Changed the photo/map container from an HBox to an HPaned.
This allows the user to shrink it in the event that he has a small screen and his photos have long filenames. The photo list will autosize (auto-grow when a long filename is added, and auto-shrink when it's removed), right up until the user resizes it for the first time. After that, it will respect their chosen size regardless of the length of filenames opened or closed until the application quits.