Change time_fudge Gtk.HScale for Gtk.SpinButton
The HScale was behaving bizarrely. Mouse wheel up caused the number to go down, and mouse wheel down caused it to go up. If I inverted it, the mouse wheel would make the numbers go the way you expect, but the slider would be visually backwards (with positive numbers on the left and negative on the right). So I dumped that mess for a SpinButton. It's a tad uglier, but it gives the user the benefit of being able to type in a number, and mouse wheel scrolling makes the numbers go the way you'd expect.