Changed time_fudge range from +/- 5 mins to +/- 1 hour.
The smaller range was chosen previously because it was difficult to mouse an accurate value from such a large range with a Gtk.HScale. Now that it's a Gtk.SpinButton, we can be a lot more liberal with this range. If you need to fudge by more than an hour, something is wrong with your timezone.