Initial work on saving EXIF data with pyexiv2
WARNING! This is a highly experimental commit and is the first bit of code in this entire project that is capable of modifying your data! So far it is only tested on JPEGs and it seems to work, but it would be INCREDIBLY STUPID to use this code on your files without making backups first. If this breaks your computer, you get to keep the pieces.

KNOWN ISSUES: There is a precision error caused by rounding. It's minor, but it needs to be fixed. Also, it doesn't save any altitude data even though we successfully extract that information from the GPX. Untested on anything but JPEGs. Please make backups before testing.