Reduce the number of calls to self.update_button_sensitivity()
I was calling this at the end of self.auto_timestamp_comparison, which was very slow because it initiates a self.liststore.foreach on each run, so every single time a new photo was added, we were iterating over the entire liststore twice. very slow. Instead, we now call it at the _end_ of the foreach only.