Major method rename and reorganization.
Moved methods around into logical groupings, renamed many methods to have more descriptive names. This commit does not change any functionality, and does not change the contents of any methods. Only their names, and the locations of the method definitions. This forms the foundation for a larger structural overhaul to come.