Unify all calls to GtkWidget.set_sensitive() to within self.update_sensitivity().
self.update_sensitivity() now ensures that all widgets have the correct sensitivity for the current state each time it is called, and is called from every place that is capable of changing program state. This introduces some redundancies, but ensures that the program never has an incorrect sensitivity on any of its widgets.