Obliterate any_modified() ugliness.
This was accomplished by implementing self.modified (a dictionary), adding to it when files are modified, and deleting from it when they are saved/reverted/closed. Essentially I've replaced a whole big slow liststore.foreach() with the builtin len(dictionary), which is an important win because it was being called almost any time the user interacted with anything at all. any_modified() has been reduced to any_modified_in_selection(), which is much less of a problem, because the selection is most likely to only have a couple things in it (or even be empty), so even with a lot of photos loaded it's not very inefficient at all.

This also allowed me to remove the self.PHOTO_MODIFIED column from the liststore, simplifying that.