Eliminate points.sort() from auto_timestamp_comparison()
Accomplished this by having the GPX parser identify the highest and lowest timestamp (self.HIGHEST and self.LOWEST), alleviating the need to copy self.tracks keys into a new list and then sort it, every single time. This is a big win because auto_timestamp_comparison() is called not just each time a photo is loaded, but it's called once for every photo each time any GPX is loaded, AND again every time the time_fudge value is changed. Considering that it's quite easy to have thousands upon thousands of points loaded from a GPX file, this is a _lot_ of sorting that is no longer happening. A small speed penalty in GPX loading has gained us a huge speed boost everywhere else.