Replace self.map_gpx (a single ChamplainPolygon) with self.polygons, a list of ChamplainPolygons. Closes #14
This allows loading of discontiguous GPX files without drawing an erroneous line connecting them, which was especially weird because the line would be from the last point of the first loaded file to the first point of the next loaded file, even if the second file that was loaded was from further into the past than the first file (ie, we were drawing lines from the most recent point to the least recent point). Note also that this commit creates a new polygon not just for each file but for each track segment within each file, allowing you to have discontiguous tracks within the same file. This can happen (at least on my GPS unit) if you turn it on, record some track points, turn it off, move somewhere else, and then turn it back on later. Track points from each day are stored in the same GPX file, but a new track segment is created each time the unit is turned on.