UI change: buttons at the bottom of the GtkListStore
Had a user complain about the apply button being immediately adjacent to the revert button and how it's easy to miss. My solution is to create a new button that toggles the photo selection (ie, if there's no selection it selects all, and if there is a selection it selects none), and then place the apply button next to it at the bottom of the liststore. I like this because it makes it really easy to manually tag all photos in two nearby clicks. I think it'll make a lot of sense for people who don't have GPS units and are bulk-tagging manually, to click 'select all' and then 'apply' one right after the other.