Replaced minidom with expat. Closes #13, #5
This had resulted in a tremendous speedup. I'll do some more thorough benchmarks later, but minidom took roughly a minute or so to load the included demo data, where expat does it literally instantly. The largest GPX file that I even have is 4.4MBs and that loads in 14s, so this is a major win.