Force map view to zoom in during GPX loading.
Previously, if you were zoomed in too far and loaded a GPX that was outside your view, the view would zoom out to accomodate the full GPX. But if you were zoomed out too far, the map view wouldn't zoom in, and if you loaded a particularly small track, it was entirely possible that you wouldn't be able to see it if you were zoomed out too far. So now we zoom in to the maximum possible level before calling ensure_visible, so that it is always forced to zoom us out until the view is Just Right, every time. Note that the back button still works as always, so you never have to be stuck with the view that ensure_visible chooses for you.