Finesse the arrow key map view movement.
The formula I'm using now is 10 / (1 + 2zoom^3). At zoom level 0 this results in 10 / (1 + 0), or "the map view moves by 10 degrees." Zoom levels range from 0 to 20, and the higher it is, the smaller this number gets, meaning the less the map view moves when you're zoomed in closer (after some quick testing it seems to be roughly half a city block in my neighborhood at the maximum zoom, which seems reasonable to me).