Revise invalid file handling in both load_exif_from_file and load_gpx_from_file. Closes #6.
This commit stops using regular expressions to make sure that a GPX file has been handed to the exif loader. Instead, what happens now is that instead of catching the RuntimeError when using GdkPixbuf to load the image, we catch that and then raise IOError (the same exception raised by pyexiv2 on invalid photos). So now for example if you try to load a GPX or an XMP or whatever else that isn't an image, load_exif_from_file raises IOError, which gets caught and then load_gpx_from_file gets the file. Also, load_gpx_from_file will now raise expat.ExpatError if the root element of the document is not 'gpx'. So it safely ignores invalid XML files.

Also, load_gpx_data now times how long it takes to parse the file, and displays the time count in the statusbar when loading completes.