Change default value of self.current['lowest'].
Changed the default value from None to "" (empty string). This is a useful trick because int < "" always evaluates True, so it's immediately clobbered by a real value, and we no longer have to test for "lowest is None" (a test which was necessary because int < None is False for some reason).