Keep consistent width of preview widget in FileChooserDialog.
Before, I'd hide the preview widget if there was no preview to show, but create a very wide preview widget if there was a preview to show. This caused the width of the file list to vary wildly depending on if there was a preview to show or not. Now, I create one single fixed-width preview widget that is always displayed, and then populate it either with an image or a 'missing image' icon as appropriate for the selected file. Also shrunk the size of the preview a bit to make more room for the file list.