Link coordinates to Google Maps. Closes #32.
The map center coordinates as well as the coordinates displayed in the FileChooserDialog for previously-tagged photos now link to Google Maps. Unfortunately the CellRendererText used in the ListStore doesn't seem to support this, so I suppressed it there. Easy workaround: click the photo you want in the ListStore, the map automatically centers on it. Then click the link centered beneath the map.