Divorce various bits of GPX parser state from GPX metadata.
New instance variable self.gpx_state is a dictionary that contains ONLY data used by the GPX parser during parsing, and is cleared when parsing is completed. self.current now only contains keys that are defined by the program (offset, highest, lowest, area, and element-name). This prevents malicious GPX files from clobbering important metadata about the GPX data. Not that that was a particularly large problem, but this is more correct.