Show some ToolButton labels in both-horiz.
So for a while now I've had my toolbar style set to "both", not realizing that the GNOME default is both-horiz and there's no real UI for changing that anymore, meaning that basically nobody using my program could see the ToolButton labels I was specifying. This rectifies that and makes my app a respectable both-horiz application.