Split gpx_element_start() into gpx_element_start() and new gpx_element_root().
gpx_element_root() is a StartElementHandler that checks that the document is a gpx before passing off the rest of the file to gpx_element_start(). The old gpx_element_start() was calling a test that was only ever true on the root element, so this commit eliminates that test from being run many thousands of times when it's never true more than once.

I'm disappointed that this did not result in a speed improvement that I was able to measure, however the resulting code is cleaner so I'm going to keep this change.