Implement self.gconf_set() and self.gconf_get().
These are simple convenience methods for accessing GConf, and will allow me to expand my use of GConf in the future, for example if I decide to implement persistent bookmarks or something else that requires GConf.

Also, this commit expands the use of APPNAME, significantly reducing the repetition of the program name, making it easier if I ever decide to rename the program later on.