Today I learned about the conditional expression.
Y'know, I have a background in perl programming, and one of the things I missed the most when I started hacking Python was that there just didn't seem to be any easy way to say 'foo = test ? biz : baz'. 'if test: foo = biz; else: foo = baz' just felt so sloppy to me. So how is it that I didn't findout about 'biz if test else baz' until just now? It's pretty ridiculous, really. At any rate, I quite like this construct. It's punctuation-free, reads fairly well, and I've decided that the conditional being in the middle is more 'endearingly quirky' than it is 'horrendously illogical'.

Also, this commit introduces the use of boolean list indices, another quirky Python feature. Maybe I'm the only person on the entire planet who didn't know this, but if you pass True as a list index, it's the same as list index 1, and False is list index 0. This greatly simplifies a few things I was doing in a similar but much more awkward way.

Oh, and, I renamed most of the keys in self.metadata. No real reason, but mostly my OCD demanded I discover synonyms for 'first' and 'last' that were the same length, so that they'd line up better, and I settled on 'alpha' and 'omega', respectively.

Cheers!