No longer using semi-private methods.
It uglied up the code and I didn't really see the point of it anyway.

Also, fairly major overhaul of the pretty string methods. Turns out I'm a huge fan of methods that consist only of a single line of input validation followed immediately by all the method logic contained in a single return statement. Got rid of a lot of single-use variables to accomplish that.