Replace self.PHOTO_* with self.col dictionary.
Having this data as a dictionary rather than distinct integer variables allows me to iterate over it in useful ways.

Considering that the values are only 0..n, I considered just having a list instead of a dictionary, but I decided against this because I need to refer to it frequently, and writing self.col['Path'] is significantly more readable than self.col.index('Path').