Massive overhaul of internal data structures.
I got fed up with GtkListStore being really fussy and not Pythonic, so I created a new data structure centered around a new object, which I've called Photograph. Photograph is essentially just a dictionary, with the exception that I'm using it's own internal namespace as the dictionary. What this means is that instead of writing photograph['attribute'], you can write photograph.attribute. Photograph provides a set() method that takes a dictionary as an argument and updates that into it's own namespace in a generic and arbitrary way, allowing me to make up any attributes I like as I go along, without being tied to any particular structure. The __init__() method creates all the attributes that I'm using at the moment, but the key victory here is that this is really easy to extend in the future if I ever need to change this, and I'm free of the shackles of the clunky GtkListStore.

Of course, the GtkListStore is still there, but it's use has been slashed. It now only stores the bare minimum needed to display the list of photos to the user: thumbnail, pango summary, filename (for matching against Photograph objects), and timestamp (for sorting).

This has greatly simplified a lot of problems I was having, such as not being able to delete ChamplainMarkers when photos were being reverted, and it also unifies a lot of the scattered support structures I had built around the ListStore, like self.has_manual, self.modified, and self.loaded_photo_iters.