Implement class ReadableDictionary.
Just a simple object that acts like a dictionary, but uses it's internal namespace as the dictionary. This allows nice things like this to work:

>>> readable = ReadableDictionary()
>>> readable['foo'] = 42
>>> readable.foo
42

It's a bit abstract, but basically it just makes dictionary accesses more readable by removing the punctuation associated with strings.

The Photograph class inherits from this, but both of these classes are used separately in my program.