Implement self.selected dictionary, removed marker_iter() and apply_single_photo().
Fairly major code cleanup. self.selected is a big win, because it frees us from TreeSelection.selected_foreach(), allowing me to remove the two methods that existed exclusively for use in selected_foreach(). It greatly simplifies all the code responsible for iterating over the selection, and it was basically free because the code that maintains self.selected was placed into update_all_marker_highlights, which was iterating over all photos every time the selection changed already anyway.

This commit also contains a few cases of minor code style cleanup where I rearranged the order of some statements so they looked better, without having any logical impact on program operation.

I also gave up on the idea of having the Revert button only be sensitive when there are modified photos in the selection, making it sensitive simply whenever there is a selection. The code for that was just so awful, removing it was a great simplification of update_sensitivity(), a method notorious for it's inefficiency. Effectively this allows users to revert unmodified files, which should be harmless.