Iterate over self.modified.keys() rather than self.modified.copy().
In save_all_files, we actually want to remove from self.modified as we iterate over it. Since you can't do that directly, I was iterating over a copy, but I realized the copy() is inefficient since I don't care about the values (which are all only True anyway), so instead I just iterate over the keys(), which is all that's needed.