Running as fast as I can just to stay in one place.
So here I am minding my own business when Fedora tells me there are software updates available. "Swell!" I think to myself, because I'm actually waiting for a radeon driver update to unbreak one of my screens. I skim over the list of updates to be installed and I noticed that pyexiv2 is to be upgraded to 0.2.2. "Curious," I think, because I know that 0.2.2 has some pretty major API breakage compared to the version that previously shipped with Fedora 14.

I install it anyway, thinking the breakage can't be that bad, but boy was I wrong. This commit features a nearly-complete rewrite of the pyexiv2 code. The best part is, pyexiv2.Rational is still completely garbage, and pyexiv2 is now even more fussy about not wanting to accept a fractions.Fraction in place of pyexiv2.Rational. So I've brought back the pug-fugly float_to_rational() method from forever ago to deal with this.

Fedora people consider this an acceptable level of API change for a "stable" release. God, I miss Debian sometimes.