Ditch create_polygon().
This method was only used in one place, so I've simply moved the code there and dropped the method definition. I've also stopped explicitely setting things that are default values for ChamplainPolygon, further simplifying the polygon creation.

Theoretically this should make the GPX parser a tiny bit more efficient, but it's probably not a measurable amount because only a handful of polygons get created during GPX parsing. It's not the kind of thing that happens during every single trackpoint, like when I ditched time.strptime and gained a noticable amount of speed.