Readability improvement in save_all_files().
Previously, I was trying to shoehorn the assignment of altitude, latitude, and longitude into a for loop, in the hopes that those things were similar enough that looping over them would simplify the code. Unfortunately, that was not the case. The assignment of Altitude was different enough from the other two that I had to create a helper method exify() to do this process, and the for loop became a mangled, unreadable mess of obscurity.

This commit unrolls the loop, ditches exify(), and I think you'll find results in a great improvement in readability.