Change how map coords are displayed. Closes #34.
I've unified position_actors() and display_coords() into one method, display_actors(). This allows the method to correct the size and positioning of self.coords_background when the width of self.coords changes.

I've added a provision in display_actors() that allows it to only update the position of the crosshair when called with no arguments, which is used during the opening animation.