Initial work on reverse geocoding support. Closes #35.
This commit introduces reverse geocoding using the JSON web services provided by GeoNames.org. Now, when you apply geotags to a photo (either automatically or manually), my program sends the coordinates to geonames.org and asks what city/state/country is located there. That data (if any) is recorded into the IPTC metadata of the photo when the 'Save All' button is clicked.