Massive internal restructuring.
Two major changes here. The first is that the Photograph object has expanded to include it's own pretty string methods, obviating the need for these within the main GottenGeography object itself. It makes more sense to have these here because now you can call these methods with no arguments and they automatically return the values associated with the Photograph instance that you've called the method on. This also simplifies how it does it's validation. Most of these methods have now been reduced to a single return statement thanks to this.

Secondly, load_exif_from_file now creates and populates the Photograph object directly, which greatly simplifies the movement of data through the load_exif_from_file() -> add_or_reload_photo() -> insert_coordinates() chain of methods. In fact insert_coordinates() has been completely removed and replaced with a much simpler modify_coordinates(). This new method is only used when _modifying_ geotag data, not when loading or reverting, so all the logic about whether or not the inserted data is 'new' or 'modified' has gone, the whole method can safely assume that the data is modified, by definition, because it modifies it. So the new method is shorter and has many fewer tests.

Of course these are sweeping changes, and vast amounts of code has been updated to work with the new stuff, as well the test suite has been updated to reflect these changes.

Also, I've added a mechanism to prevent re-downloading geonames.org data each time the clock offset changes, because tha'ts fantastically slow and is typically used for minor fine-tuning (eg, changing the clock offset is not likely to land you in another city, so it just keeps the data it had previously rather than redownload it each time, as was the previous inefficient behavior.