Ditch blocking urllib2 for asynchronous Gio methods.
This commit introduces a pretty major win in terms of application responsiveness. Using Gio instead of urllib2 to load geonames.org webservices is both faster, and asynchronous, which allows the user to continue using the program while we load the reverse geocode data in the background.

Users of slow internet connections may notice a delay between when they tag their photos and when the reverse geocode data shows up, but on broadband it's almost unnoticable.

Also, I've introduced a rate limiter on the geonames request. No more than 1 request per photo per 10 seconds. The problem was that using the mousewheel to scroll the clock offset setting generated a vast number of requests that were unlikely to have different results (eg, the clock offset is for subtle corrections. In order for a tweak to the clock offset setting to produce a different reverse geoname lookup, you'd have to have been moving so fast that a couple minutes got you from one city to the next). This rate limitation also applies to manual tags, so if you manually tag a photo into the wrong city, you'll have to wait 10 seconds before you hit apply again if you want to get updated geocode data. If you're just making subtle adjustments within the same city, this won't matter to you.