Implement queueing for reverse geocode requests.
Let's say hypothetically you have 100 photos taken within a few meters of each other. The previous behavior was to send out a webservice request for the geocode data for each photo, and each response would then be cached. Subsequent requests would be drawn from the cache (for example, if you manually tweaked the location of a couple photos, or if you were adjusting the clock offset setting). But the initial photo loading would result in 100 identical (or nearly identical) requests sent to geonames.org.

This commit changes this behavior so that each geonames.org request can check if a request has already been sent (even if the response hasn't returned yet), and if so, add the photo to a queue so that it can get the geocode data when the response does return. This way, we never request the same (or similar) location twice. This has resulted in such a big increase in efficiency, that I've dropped the rate limiter altogether.

The cache caches the results with 0.01 degree precision, which is roughly one square kilometer. What this means is that if you are making precise adjustments to the location of your photo (say, just a few meters), GottenGeography won't send another request for geocode data at all. You have to move about 1km away from the original location in order to generate a new request. This cache is in memory only, so if you want to force a re-request, just close and reopen the program.