Stage 1 of a major structural reorganization.
Gottengeography thus far has been developed as one large monolithic application. That was ok for a while, but it's becoming unmanageably large. I'm now breaking up the application into a python package with multiple modules that should greatly increase the maintainability of the code.

So far just the easy bits have been extracted, and (importantly) the test suite still passes.