Stage 5 of a major structural reorganization.
This commit introduces the GeoCache class, which serves both as a data store for caching geonames.org webservice data, and as a go-between interface between the Photograph class and the GottenGeography class. The important distinction here is that GottenGeography no longer passes itself as an argument to the Photograph class, meaning that Photograph methods are no longer able to access the full API inside GottenGeography, cleaning up that interface significantly. I've also changed the GPXLoader class so that it does not require the GottenGeography instance be passed to it, either.

Also, modify_coordinates and set_marker_highlight methods made more sense as Photograph instance methods, and so have been moved there.