Implement color picker widget in preferences dialog.
The color picker widget can now be used to control the color of GPX tracks. It's instant-apply, and the results can be seen immediately (or they could, if libchamplain would repaint itself properly like it should, but that's another bug). The track color also persists with GConf.

TODO: test what happens with the track color on first run (I'm guessing it'll default to black, which I should fix), and implement the "cancel" button properly so that it actually undoes the changes.