Connect timezone prefs to the timezone functionality.
It's finally done! This is pretty sweet. With a single click, my application can lookup what timezone the photos were taken in and then compute the locations correctly. No more tinkering with the clock offset settings if your camera's and computer's clocks aren't syncronized!

The only issue is that geonames.org is becoming increasingly unreliable due to overloaded servers. I'll have to download their db and start accessing it locally rather than as a web request. No rest for the wicked...