Rip out geonames.org webservice entirely.
I've downloaded the relevant parts of their database and am now processing it locally. This is not only faster, it's more reliable, and it makes the timezone-lookup feature feasible.

The GeoCache object no longer queues requests for callback because this approach is so vastly faster that it just doesn't matter, however since this also changes from an asynchronous lookup back to a synchronous one, I've kept the cache just in case you have a slow disk or something. My machine does the reverse geocode lookup in 0.35 seconds on average, and once it's cached in the GeoCache object, it's instantaneous for other nearby photos.