Use cPickle to persist the track color RGB values rather than having 3 separate keys for each of red, green blue.
The benefit here is that it's easier to distinguish whether the user has selected black (0,0,0) as their track color or if the program has never run before and the pickled data is therefore missing, making it very easy to specify the desired default first-run color, which I've chosen to be purple.

Also, pickling is awesome.