Fix several bugs with the map search feature.
No longer loading & displaying duplicate results if you type the same three letters to initiate two separate searches.

No longer loading extraneous cities (eg, not loading "Viedma" as a search result for "edm". This is a fair performance enhancement. Note also that you *can* get search results for the not-first word in the name of a city. For example, Typing in "joh" returns both 'Johannesburg' and 'Saint John', but not 'Hajoho'.