Change the way modules are imported.
I've changed a lot of instances of "import foo" to "from foo import bar". This has a few effects. First, calling bar() instead of foo.bar() is known to be faster. Second, it simplifies the look of a lot of the code, shortening lines.

A downside to this is that everything imported (for example) in gps.py is then also imported into datatypes.py with the line "from gps import *". So, I've implemented the special __all__ variable to determine what * imports and not to import anything extraneous. This greatly cleans up the namespace of each module, preventing anything unneeded from being imported.