Replace a slow for loop with much faster list comprehensions.
This is twice as fast as the old behavior, according to the timeit module, even though two separate list comprehensions means that technically we're iterating over self.tracks twice instead of once.

Here's the old behavior condensed into timeit form (t represents self.tracks with some dummy data, s represents timestamp with a value not in t, hi and lo are the first and last values of t but are then computed to be the two nearest values to s):

$ python -mtimeit -s "t=range(1000000000,1000002000,2); s=1000001001; hi=t[-1]; lo=t[0]" "for p in t:" "  if p > s: hi = min(p, hi)" "  if p < s: lo = max(p, lo)"
1000 loops, best of 3: 304 usec per loop

And here's the new behavior:

$ python -mtimeit -s "t=range(1000000000,1000002000,2); s=1000001001; hi=t[-1]; lo=t[0]" "hi=min([p for p in t if p > s]);lo=max([p for p in t if p < s])"
10000 loops, best of 3: 186 usec per loop

So that's a speedup of 60%. It's counterintuitive that iterating over the same list twice would take half the time of doing it once. The reason for this is that python has a certain amount of overhead associated with for loops. In the previous behavior, the for loop would execute the "min" and "max" builtins for every single element of the list. Each function invocation also has a certain amount of overhead associated with it.

By using list comprehensions instead, the iteration is done in much faster C code, and instead of invoking min() and max() once for every single iteration, it just has one simple test to evaluate. Then the comprehension returns a simple list to each of min() and max() which are then invoked only once, which again go into fast C code to choose the highest and lowest value much more quickly than we could hope to accomplish in python code. Indeed, each list comprehension only takes 92 microseconds:

$  python -mtimeit -s "t=range(1000000000,1000002000,2); s=1000001001; hi=t[-1]; lo=t[0]" "hi=min([p for p in t if p > s])"
10000 loops, best of 3: 92 usec per loop