Simplify GPXLoader API.
Instead of taking in a bunch of objects and trying to manipulate them directly, it now only takes three arguments: the filename, and two callbacks.

The first callback is called every 0.2s during parsing, in order to update the display. It gives an opportunity to pulse a progressbar and redraw the app so it doesn't look hung.

The second callback is called whenever a new ChamplainPolygon is required. This callback creates the polygon, decorates it however is desired, adds it to the map, and then returns the polygon.

This change means that GPXLoader doesn't know anything at all about Gtk or libchamplain. In theory, if somebody else developed a Polygon class that was compatible with Champlain.Polygon (all that's necessary is a method like polygon.append_point(lat, lon)), then it is possible to use GPXLoader in a different application altogether, without modification.

For example, if you were to develop a commandline app that needed to parse GPX data, you'd just have to make an object with the append_point method, and a function that instantiates that object and returns it, and another method could print len(gpx.tracks) to display how many tracks have been loaded each 0.2s.