auto_timestamp_comparison() did not need to be an instance method on GottenGeography class.
This method gets called fairly frequently, so performance is critical. By making it a top level method and passing arguments to it (as opposed to accessing attributes of 'self'), it becomes more efficient. This is because arguments passed to the function appear in the function local namespace, which is searched first when trying to find the value of a variable. By having to lookup self.tracks, it adds an extra step before it can find the value it needs.