Simplify MarkerController.update_highlights().
Some of the complexity here is leftover from when this method used to do more, but now I've been able to streamline it a bit.

Also this makes a minor, unrelated change to SearchController.match_func, making it slightly more efficient by passing it a method that it calls rather than passing it an object and having it call the method on the object. One less symbol table lookup in a function that's called frequently during searches.