Add the ability to specify files to open on the commandline.
Typically paths on the commandline will be specified relative to the current working dir, while paths from GtkFileChooser are absolute. This resulted in a problem where if you used both of those methods to load the same file, GottenGeography wouldn't realize that they were the same file. To avoid this situation I run all the files through os.path.abspath so that the paths are stored internally the same way regardless of how you open them. This eliminates the possibility of duplicate loading.