Change google maps link text to be a tooltip, with the visible text saying just 'Google'.
The problem with this link is that it easily results in the window being widened to accomodate it, which is worsened by the fact that translations can easily input a longer string here than necessary. By making the text a tooltip, it doesn't affec the width of the window, and translators can insert whatever text they like here.