Stop remembering location as a signal handler on notify::(latitude|longitude).
Was having lots of trouble with the threshhold before, and it wasn't always remembering when I wanted it to. This commit reverts it back to the old behavior, of having to call remember_location() everywhere that changes the map view, which is admittedly error-prone, but seems to just work a whole lot better as long as you remember to call it from everywhere necessary.