Connect NavigationController.remember_location() to ChamplainView::realize signal.
This probably doesn't make sense at first glance, and it kind of doesn't. I'm connecting to 'realize' because this is a very infrequently-used signal (basically once at the start, and that's it). By connecting to this signal, I then have the ability to call self.map_view.emit("realize") whenever I want the current location to be remembered. This alleviates the need to pass remember_location as an argument to the other Controller classes, because they can call it indirectly by emitting the signal, rather than knowing anything about the method and calling it directly.

I've asked Jiri Techet to implement a better signal for this purpose that I can use unambiguously, but this is working well for now.