Snap back to the last-searched location by pressing Enter in the search bar.
This is a little bit subtle, but I think it's an important feature. Before this, if you searched for a location, the only way to make the map actual visit your selected location was to select it from the drop-down list of results (either by mouse or keyboard was fine). If you then navigated away from that place and wanted to return to it, you had to delete at least one character from your search string, causing the GtkEntryCompletion to reappear, so you could reselect the location to jump to. Now, you can simply click in the search bar, and hit enter, and it automatically jumps back to the previous location, no problems.

The only problem I foresee with this is that if you start typing a new search and hit enter without selecting something from the GtkEntryCompletion first, it'll go back to the last search you did. There's potential for confusion here, but I think it's a good start. Perhaps some kind of test that the entry text matches the previous search would be fine, but I'll wait until somebody actually complains about this behavior before I fix it.