Abolish GottenGeography.listsel in favor of LabelController.selection.
They both point to the exact same GtkTreeSelection object, it was redundant having two references to it. The new code is a little bit deceptive because 'self.labels.selected' has nothing to do with a ChamplainLabel or any of the new Champlain selection logic, however the selection does directly map (ie, there's a 1:1 mapping of photos to labels, and the selection identifies which photos are selected, so by proxy it also represents which labels are selected).